requirements.txt
torch>=2.10
torchvision>=0.25
timm>=1.0
albumentations>=2.0
Pillow>=12.0
numpy>=2.0
opencv-python>=4.10
transformers>=5.0
tqdm
matplotlib
scikit-learn
